pyproject.toml
[project]
name = "home-catalog"
version = "0.1.0"
description = "Catalog of goods for home"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "daphne==4.1.2",
    "django==5.0.4",
    "django-filter>=24.3",
    "djangorestframework>=3.15.2",
    "social-auth-app-django==5.4.2",
    "whitenoise==6.6.0",
    "python-slugify>=8.0.4",
]
[project.optional-dependencies]
  develop = [
    "esbuild-py @ git+https://github.com/ErDmKo/esbuild-py@add_build_api",
    "ansible>=10.7.0",
    "ruff==0.4.4",
    "coverage>=7.4.0",
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.1.0",
  ]
